Code Updations on cb-course-aggregator Flink Job
