+++ data-pipeline-flink/cb-course-aggregator/src/main/resources/cb-course-aggregator.conf
@@ -3,7 +3,7 @@ include "base-config.conf"
 kafka {
   output.topic = ${job.env}".assessment.final"
   assessment.input.topic = ${job.env}".assessment.submit"
-  generate.certificate.input.topic = ${job.env}".generate.certificate.request"
+  issue.certificate.input.topic = ${job.env}".issue.certificate.request"
   groupId = ${job.env}"-assessment-submit-group"
   success.topic = ${job.env}".success.assessment.submit"
 }
