+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/cache/DataCache.scala
@@ -92,7 +92,7 @@ class DataCache(val config: BaseJobConfig, val redisConnect: RedisConnect, val d
     val data = redisConnection.get(key)
     if (data != null && !data.isEmpty) {
       val dataMap = gson.fromJson(data, new util.HashMap[String, AnyRef]().getClass)
-      if (fields.nonEmpty) dataMap.keySet().retainAll(fields.asJava)
+      if(fields.nonEmpty) dataMap.keySet().retainAll(fields.asJava)
       dataMap.values().removeAll(util.Collections.singleton(""))
       dataMap.asScala
     } else {
